# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fuyao Procurement Cloud API
  version: 0.1.0
  description: >
    Backend contract for the Fuyao procurement SaaS. This mirrors exactly what
    the frontend calls in `src/lib/api.ts`; the data shapes come from
    `src/types.ts` (the single source of truth). Implement this and the
    frontend works with no changes — just set `VITE_API_BASE_URL` to your URL.


    Conventions:
      * REST + JSON, all paths under `/api`.
      * Auth: every endpoint except `POST /api/auth/login` requires
        `Authorization: Bearer <JWT>`.
      * Errors use the envelope `{ "error": { "code", "message" } }`.
      * Timestamps are epoch MILLISECONDS (numbers), matching JS `Date.now()`.
        If you store `timestamptz` in Postgres, convert to ms on the way out.
      * `id` fields are strings (UUID recommended).
servers:
  - url: '{baseUrl}'
    description: Backend base URL (set as VITE_API_BASE_URL in the frontend)
    variables:
      baseUrl:
        default: http://localhost:8000
security:
  - bearerAuth: []
...
paths:
...
  /api/sourcing/search:
    post:
      tags: [Sourcing]
      summary: Find suppliers for a natural-language requirement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourcingSearchRequest'
      responses:
        '200':
          description: Matching suppliers
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Supplier'
  /api/sourcing/search-jobs:
    post:
      tags: [Sourcing]
      summary: Start an asynchronous supplier research job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourcingSearchRequest'
      responses:
        '202':
          description: Search job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchJobResponse'
  /api/sourcing/search-jobs/{job_id}:
    get:
      tags: [Sourcing]
      summary: Poll an asynchronous supplier research job
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Current job state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchJobResponse'
        '404':
          $ref: '#/components/responses/Error'
  /api/sourcing/search-jobs/{job_id}/events:
    get:
      tags: [Sourcing]
      summary: Stream supplier research job progress as Server-Sent Events
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Server-Sent Events containing SearchJobResponse payloads
          content:
            text/event-stream:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: INVALID_CREDENTIALS
            message:
              type: string
              example: Email or password is incorrect
...
    SourcingSearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          example: Water deflector strips and glass adhesive in Germany
        structured:
          $ref: '#/components/schemas/SourcingStructuredFields'

      type: object
      description: Optional structured sourcing fields; they override/augment parsed intent.
      properties:
        productName: { type: string }
        quantity: { type: string }
        unit: { type: string }
        brand: { type: string }
        category: { type: string }
        country: { type: string }
        certifications: { type: string }
    SupplierCategory:
      type: string
      description: Open category string. Known examples include waterDeflector, glassAdhesive, rubberSeal, glassRaw, hardware, packaging, cleaning, office, safetyShoes, firstAid, equipment.
    Supplier:
      type: object
      required:
        [id, name, category, country, city, address, contactPerson, phone, email,
         website, employees, annualRevenue, established, capabilities, certifications, matchScore]
      properties:
        id: { type: string }
        name: { type: string }
        category: { $ref: '#/components/schemas/SupplierCategory' }
        country: { type: string }
        city: { type: string }
        address: { type: string }
        contactPerson: { type: string }
        phone: { type: string }
        email: { type: string }
        website: { type: string }
        employees: { type: string, description: 'Headcount band, e.g. "250–500"' }
        annualRevenue: { type: string, description: 'e.g. "€ 80M–120M"' }
        established: { type: integer, example: 1995 }
        capabilities:
          type: array
          items: { type: string }
        certifications:
          type: array
          items: { type: string }
        matchScore:
          type: number
          minimum: 0
          maximum: 100
          description: Relevance to the query (0–100)
    SearchJobEvent:
      type: object
      required: [timestamp, phase, message, progress]
      properties:
        timestamp: { type: integer, format: int64, description: Epoch milliseconds }
        phase: { type: string }
        message: { type: string }
        progress: { type: integer, minimum: 0, maximum: 100 }
    SearchJobResponse:
      type: object
      required: [jobId, status, progress, step, events, results]
      properties:
        jobId: { type: string }
        status:
          type: string
          enum: [queued, running, completed, failed]
        progress: { type: integer, minimum: 0, maximum: 100 }
        step: { type: string }
        events:
          type: array
          items: { $ref: '#/components/schemas/SearchJobEvent' }
        intent:
          type: object
          nullable: true
          additionalProperties: true
        results:
          type: array
          items: { $ref: '#/components/schemas/Supplier' }
        error:
          type: string
          nullable: true

      type: string
      enum: [onAccount, prepayment, card]
...